Add flexibility into API (#1076)
* remove taking damage when cable loss is == 0
add possibility to disable direct smelting
add possibility to change crushed material
SimpleGeneratorMetaTileEntity has now a createWorkableHandle
INSULATION_MATERIALS is now public to add more insulation tier

* fix getTierByVoltage never throw ArrayOutOfBoundException
Total EU is now long

+++ src/main/java/gregtech/integration/jei/recipe/GTRecipeWrapper.java
@@ -108,7 +108,7 @@ public class GTRecipeWrapper implements IRecipeWrapper {
 
     @Override
     public void drawInfo(Minecraft minecraft, int recipeWidth, int recipeHeight, int mouseX, int mouseY) {
-        minecraft.fontRenderer.drawString(I18n.format("gregtech.recipe.total", Math.abs(recipe.getEUt()) * recipe.getDuration()), 0, 70, 0x111111);
+        minecraft.fontRenderer.drawString(I18n.format("gregtech.recipe.total", Math.abs((long) recipe.getEUt()) * recipe.getDuration()), 0, 70, 0x111111);
         minecraft.fontRenderer.drawString(I18n.format(recipe.getEUt() >= 0 ? "gregtech.recipe.eu" : "gregtech.recipe.eu_inverted", Math.abs(recipe.getEUt())), 0, 80, 0x111111);
         minecraft.fontRenderer.drawString(I18n.format("gregtech.recipe.duration", recipe.getDuration() / 20f), 0, 90, 0x111111);
         int baseYPosition = 100;
